+++ scripts/generate-component-config/build-configs.js
@@ -1,0 +1,18 @@
+const terraNameSpacedRepos = [
+  'terra-core',
+  'terra-framework',
+];
+
+const blackListedDirectories = [
+  'examples',
+  'test-examples',
+  'src',
+  'tests',
+  'nightwatch',
+  'components',
+];
+
+module.exports = {
+  terraNameSpacedRepos,
+  blackListedDirectories,
+};
